// packages/monaco-aom/src/core/parser/gen/ast.ts
export type Block = AppDefBlock | CompBlock | CompDefBlock | ProviderBlock | SecretDefBlock;

export const Block = 'Block';
// ...
export interface AppDefBlock extends AstNode {
    readonly $container: Model;
    readonly $type: 'AppDefBlock';
    appBlocks: Array<AppBlock>
    name: string
}

export const AppDefBlock = 'AppDefBlock';
// ...
export interface CompBlock extends AstNode {
    readonly $container: AppDefBlock | Model;
    readonly $type: 'CompBlock';
    compBlocks: Array<CompCompBlock>
    name: string
}

export const CompBlock = 'CompBlock';

export function isCompBlock(item: unknown): item is CompBlock {
    return reflection.isInstance(item, CompBlock);
}

export interface CompDefBlock extends AstNode {
    readonly $container: Model;
    readonly $type: 'CompDefBlock';
    name: string
    props: Array<Property>
}

export const CompDefBlock = 'CompDefBlock';
// ...
export interface Model extends AstNode {
    readonly $type: 'Model';
    blocks: Array<Block>
    imports: Array<ModelImport>
}

export const Model = 'Model';

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model);
}

export interface ModelImport extends AstNode {
    readonly $container: Model;
    readonly $type: 'ModelImport';
    path: string
}

export const ModelImport = 'ModelImport';
// ...
export interface ProviderBlock extends AstNode {
    readonly $container: Model;
    readonly $type: 'ProviderBlock';
    name: string
    props: Array<Property>
}

export const ProviderBlock = 'ProviderBlock';
// ...
export interface SecretDefBlock extends AstNode {
    readonly $container: Model;
    readonly $type: 'SecretDefBlock';
    name: string
    props: Array<Property>
}

export const SecretDefBlock = 'SecretDefBlock';